edicion funcional con back

+++ app/src/main/java/com/example/primerparciallabov/Producto.java
@@ -3,13 +3,13 @@ package com.example.primerparciallabov;
 public class Producto {
     public String nombre;
     public int cantidad;
-    public double precio;
+    public float precio;
 
     public Producto(){
 
     }
 
-    public Producto(String nombre, int cantidad, double precio){
+    public Producto(String nombre, int cantidad, float precio){
         this.nombre = nombre;
         this.cantidad = cantidad;
         this.precio = precio;
